Split out cross platform test
