+++ src/integration/pages/storyPage/canonicalTests.js
@@ -7,6 +7,19 @@ export default () => {
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
 
+  describe('Lead image', () => {
+    it('I can see an image with a caption in a noscript tag', () => {
+      const imageEl = document.querySelector('main figure');
+      console.log(`here: ${JSON.stringify(imageEl, null, 2)}`);
+      expect(imageEl).toBeInTheDocument();
+
+      const imageCaptionEl = document.querySelector('main figure figcaption');
+      expect(imageCaptionEl).toBeInTheDocument();
+      expect(imageCaptionEl.textContent).toBeTruthy();
+      expect(imageCaptionEl.textContent).toMatchSnapshot();
+    });
+  });
+
   describe('Social Embeds', () => {
     const hasRichInstagramEmbed = !!document.querySelector(
       'iframe.instagram-media',
